favorite modified

+++ favorites/views.py
@@ -1,3 +1,4 @@
+from django.db import IntegrityError
 from django.http import JsonResponse
 from rest_framework import status
 from rest_framework import generics, permissions
@@ -19,14 +20,16 @@ class CreateDestroyAPIView(APIView):
     permission_classes = (permissions.IsAuthenticated,)
 
     def put(self, request, pk):
+        if not request.user.is_authenticated:
+            return JsonResponse({"detail": "You have to login"}, status=status.HTTP_403_FORBIDDEN)
         try:
             favorite = FavoritePost.objects.create(post_id=pk, user_id=request.user.id)
             favorite.save()
-            return JsonResponse({'favorite': 'created'}, status=status.HTTP_201_CREATED)
-        except:
+            return JsonResponse({'favorite': True}, status=status.HTTP_201_CREATED)
+        except IntegrityError:
             favorite = FavoritePost.objects.get(post_id=pk, user_id=request.user.id)
             favorite.delete()
-            return JsonResponse({'favorite': 'destroyed'}, status=status.HTTP_200_OK)
+            return JsonResponse({'favorite': False}, status=status.HTTP_200_OK)
 
 
 class UserInfo(APIView):
